refactor(client): organize pages
